Fixed up the jxc_cpp converter system: * Improved stdlib type handling * Added macros for quickly defining simple struct and enum converters * Added second web server example using the new converter API * Added converter tests * Some general bugfixes

+++ jxc/include/jxc/jxc_bytes.h
@@ -100,7 +100,7 @@ JXC_END_NAMESPACE(traits)
 JXC_BEGIN_NAMESPACE(detail)
 
 // Simple stack-based char buffer, generally intended for small strings, but could be used for anything
-template<typename T, size_t BufSize>
+template<typename T, uint16_t BufSize>
 struct MiniBuffer
 {
     using value_type = T;
